Feature/projects module list and element

+++ frontend/controllers/ProjectController.php
@@ -2,6 +2,7 @@
 
 namespace frontend\controllers;
 
+use common\models\ProjectsOnStagesByUser;
 use Yii;
 use common\models\ProjectsOnStages;
 use common\models\query\ProjectQuery;
@@ -35,9 +36,16 @@ class ProjectController extends Controller
             'access' => [
                 'class' => AccessControl::className(),
                 'rules' => [
+                    //сначала запрещающие правила
+                    //пример простого использования rbac, если не нужны правила (rules)
+                    /*[
+                        'allow' => false,
+                        'actions' => ['index'],
+                        'roles' => ['manager'],
+                    ],*/
                     [
                         //'actions' => [/*'logout', 'index', 'update', 'view',*/ 'create'/*, 'delete', 'my'*/],
-                        'actions' => ['logout', 'index', 'update', 'view', 'create', 'delete'],
+                        'actions' => ['logout', 'index', 'index-by-user', 'update', 'view', 'create', 'delete'],
                         'allow' => true,
                         'roles' => ['@'],
                     ],
@@ -101,6 +109,10 @@ class ProjectController extends Controller
             'query' => ProjectsOnStages::find(),
         ]);
 
+        $dataProviderProjectStagesByUser = new ActiveDataProvider([
+            'query' => ProjectsOnStagesByUser::find(),
+        ]);
+
         //провайдер для получения имени активного этапа по номеру
         $dataProviderStageTitle = new ActiveDataProvider([
             'query' => ProjectsOnStages::find()->where('fk_stage = ' . $fk_stage),
@@ -111,6 +123,57 @@ class ProjectController extends Controller
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
             'dataProviderProjectStages' => $dataProviderProjectStages,
+            'dataProviderProjectStagesByUser' => $dataProviderProjectStagesByUser,
+            'dataProviderStageTitle' => $dataProviderStageTitle,
+            'fk_stage' => $fk_stage,
+        ]);
+    }
+
+    public function actionIndexByUser($fk_stage = -1)
+    {
+        //$searchModel = new ProjectSearch();
+        //$dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+        /* @var $query ProjectQuery */
+        //$query = $dataProvider->query;
+        //$query->byUser(Yii::$app->user->id);
+        /*return $this->render('index', [
+            'searchModel' => $searchModel,
+            'dataProvider' => $dataProvider,
+        ]);*/
+
+        $searchModel = new ProjectSearch();
+        //если фильтров нет
+        if ($fk_stage == -1) {
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+        } //если стоит фильтр по этапу проекта, то отфильтровать список
+        else {
+            $dataProvider = new ActiveDataProvider([
+                'query' => ProjectModel::find()->joinWith('projectUsers')
+                    ->where('fk_stage=' . $fk_stage)->andWhere('user_id='.Yii::$app->user->id)
+            ]);
+        }
+        $dataProvider->pagination->pageSize = 12; //пагинация по 12 записей на странице
+
+        //провайдер для вывода перечня фильтров
+        $dataProviderProjectStages = new ActiveDataProvider([
+            'query' => ProjectsOnStages::find(),
+        ]);
+
+        $dataProviderProjectStagesByUser = new ActiveDataProvider([
+            'query' => ProjectsOnStagesByUser::find(),
+        ]);
+
+        //провайдер для получения имени активного этапа по номеру
+        $dataProviderStageTitle = new ActiveDataProvider([
+            'query' => ProjectsOnStages::find()->where('fk_stage = ' . $fk_stage),
+        ]);
+
+
+        return $this->render('index', [
+            'searchModel' => $searchModel,
+            'dataProvider' => $dataProvider,
+            'dataProviderProjectStages' => $dataProviderProjectStages,
+            'dataProviderProjectStagesByUser' => $dataProviderProjectStagesByUser,
             'dataProviderStageTitle' => $dataProviderStageTitle,
             'fk_stage' => $fk_stage,
         ]);
@@ -136,15 +199,18 @@ class ProjectController extends Controller
      */
     public function actionCreate()
     {
-        $model = new ProjectModel();
+        if (\Yii::$app->user->can('crudProject')) {
+            $model = new ProjectModel();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+            return $this->render('create', [
+                'model' => $model,
+            ]);
         }
-
-        return $this->render('create', [
-            'model' => $model,
-        ]);
+        return $this->redirect('/project');
     }
 
     /**
@@ -156,15 +222,18 @@ class ProjectController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if (\Yii::$app->user->can('crudProject')) {
+            $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+            return $this->render('update', [
+                'model' => $model,
+            ]);
         }
-
-        return $this->render('update', [
-            'model' => $model,
-        ]);
+        return $this->redirect('/project');
     }
 
     /**
@@ -176,9 +245,12 @@ class ProjectController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if (\Yii::$app->user->can('crudProject')) {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        }
+        return $this->redirect('/project');
     }
 
     /**
